requirements_pypi.txt
absl-py==0.15.0
astor==0.8.1
astunparse==1.6.3
black==21.10b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
flatbuffers==1.12
gast==0.4.0
google-api-core==2.2.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-vision==2.6.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
h5py==3.1.0
idna==3.3
imutils==0.5.4
keras==2.7.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mypy-extensions==0.4.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
pandas==1.3.4
pathspec==0.9.0
platformdirs==2.4.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyftdi==0.53.3
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
pyusb==1.2.1
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-docs==0.0.0.dev0
tensorflow-estimator==2.7.0
tensorflow-macos==2.6.0
termcolor==1.1.0
tomli==1.2.2
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
